+++ .bud/Build.cs
@@ -22,7 +22,8 @@ public class Build : IBuild {
                                      Cs.Dependency("Bud.NUnit.ConsoleRunner"),
                                      Cs.Dependency("CommandLineParser")),
                               Cs.Test(Cs.RootNamespace.Modify("Bud"),
-                                      Cs.Dependency("NUnit")));
+                                      Cs.Dependency("NUnit"),
+                                      Cs.Dependency("Moq")));
 
     var bud = new Project(BudProjectId,
                           Cs.Exe(Cs.RootNamespace.Modify("Bud"),