+++ Bud.Core/src/main/cs/Cli/IConsoleInput.cs
@@ -1,0 +1,7 @@
+using System;
+
+namespace Bud.Cli {
+  public interface IConsoleInput {
+    ConsoleKeyInfo ReadKey();
+  }
+}